RoomService added; LocalizationHelper added; Everything changed to use LocalizationHelper to get localized strings
RoomService was a standard implementation of the IRoomService interface though only SimpleRoomService was actually tested in the field, so I guess the Simple implementation is now standard.
LocalizationHelper added in order to be able to use two languages (the bot is using Russian primarily. The English versions of the strings are added for the better understanding of those who's going to clone this repository).